+++ R/methods-default.R
@@ -1,8 +1,10 @@
 ## Default methods for data (will be for inz.simple and anything without a method)
 
-gSubset <- function(df, g1.level, g2.level, df.vs, missing)
+gSubset <- function(df, g1.level, g2.level, df.vs, missing) {
     UseMethod("gSubset")
+}
 
+#' @exportS3Method
 gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
     # subset the data by g2 (keep everything, so xlims can be calculated)
     # g2 can take values (0 = "_ALL", 1:ng2, ng2+1 = "_MULTI")
@@ -15,8 +17,9 @@ gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
         # if g2 specified numerically, check the value is ok, and then convert it to
         # character level anyway
         if (is.numeric(g2.level)) {
-            if (as.integer(g2.level) != g2.level)
+            if (as.integer(g2.level) != g2.level) {
                 warning(paste0("g2.level truncated to ", g2.level, "."))
+            }
 
             if (g2.level == 0) {
                 g2.level <- "_ALL"
@@ -39,7 +42,8 @@ gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
             }
 
             missing$g2 <- sum(is.na(df$data$g2))
-            df1 <- lapply(g2l,
+            df1 <- lapply(
+                g2l,
                 function(l) {
                     dft <- df$data[df$data$g2 == l & !is.na(df$data$g2), , drop = FALSE]
                     dft[, colnames(dft) != "g2"]
@@ -60,7 +64,7 @@ gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
     if ("g1" %in% df.vs) {
         # take two methods of specifying g1.level (numeric or level names), and convert to a vector
         # of only character names to be plotted
-        g1l <- levels(df$data$g1)  # all levels of variable
+        g1l <- levels(df$data$g1) # all levels of variable
         if (is.null(g1.level)) g1.level <- "_MULTI"
 
         if (is.numeric(g1.level)) {
@@ -68,8 +72,9 @@ gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
             g1.level <- if (any(g1.level == 0)) "_MULTI" else levels(df$data$g1)[g1.level]
         }
 
-        if (any(g1.level == "_MULTI"))
+        if (any(g1.level == "_MULTI")) {
             g1.level <- levels(df$data$g1)
+        }
 
         # track missing values due to missingness in g1
         missing$g1 <- sum(is.na(df$data$g1))
@@ -80,9 +85,11 @@ gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
 
     # this converts each data.frame in the list to a list of data
     # frames for all levels of g1
-    df.list <- lapply(df1,
+    df.list <- lapply(
+        df1,
         function(df2) {
-            df3 <- lapply(g1l,
+            df3 <- lapply(
+                g1l,
                 function(x) inzDataList(df2, x)
             )
             names(df3) <- g1l
@@ -92,25 +99,36 @@ gSubset.default <- function(df, g1.level, g2.level, df.vs, missing) {
 
     ## sum up all of the missing values
     w.df <-
-        if (is.null(g2.level)) "all"
-        else if (g2.level == "_MULTI") 1:length(df.list)
-        else g2.level
+        if (is.null(g2.level)) {
+            "all"
+        } else if (g2.level == "_MULTI") {
+            1:length(df.list)
+        } else {
+            g2.level
+        }
 
     missing$x <- sum(
-        sapply(df.list[w.df],
-            function(df) sum(
-                sapply(df, function(d) sum(is.na(d$x)))
-            )
+        sapply(
+            df.list[w.df],
+            function(df) {
+                sum(
+                    sapply(df, function(d) sum(is.na(d$x)))
+                )
+            }
         )
     )
-    if ("y" %in% df.vs)
+    if ("y" %in% df.vs) {
         missing$y <- sum(
-            sapply(df.list[w.df],
-                function(df) sum(
-                    sapply(df, function(d) sum(is.na(d$y)))
-                )
+            sapply(
+                df.list[w.df],
+                function(df) {
+                    sum(
+                        sapply(df, function(d) sum(is.na(d$y)))
+                    )
+                }
             )
         )
+    }
 
     class(df.list) <- "inz.simple"